requirements.txt
Flask
Flask-CORS
python-docx
pdfminer.six
google-generativeai
python-dotenv
